goal_agent/requirements.txt
crewai>=0.1.0
openai>=1.0.0
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0.0
requests>=2.28
tenacity>=8.0.0
pydantic>=1.10
python-dateutil>=2.8
boto3>=1.26.0
